Keep script name as main.py for development
