+++ docker/Dockerfile
@@ -8,5 +8,5 @@ RUN pip3 install --no-cache-dir -r requirements.txt
 
 COPY ../src /app
 
-#TODO: replace "main.py" with actual script file name when script is done
+#TODO: csv_to_sql.py with final version that uses firestore as datasource
 CMD ["python", "main.py"]
