+++ docker/requirements.txt
@@ -1,2 +1,5 @@
 # TODO: Replace with actual requirements when python script is ready
 google-cloud-firestore==2.21.0
+pandas
+sqlalchemy
+numpy
