--- src/requirements.txt
@@ -1,3 +1,0 @@
-pandas
-sqlalchemy
-numpy
